Issue #250 Compile for projects targeting dot.net 4.7.2
Dependant projects targeting dot.net 4.7.2 and later have runtime errors due to Tuple mismatches between the dot.net SDK and the Antlr4 library, when using the package from nuget.  If the compiler project is updated to target 4.7.2 then the runtime errors are caught during compilation.  Updated the project to demonstrate the problem during build, and updated the code to pass the tests.

Note: this is backwards incompatible, due to the change in the targeted framework.